% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.calcAtSeaObserverEffort.R
\name{adfg.calcAtSeaObserverEffort}
\alias{adfg.calcAtSeaObserverEffort}
\title{Calculate at-sea observer sampling effort}
\usage{
adfg.calcAtSeaObserverEffort(tbl)
}
\arguments{
\item{tbl}{- measure pot data table from \code{adfg.extractMPD}}
}
\value{
dataframe with columns
* fishery
* area
* year
* numpots
}
\description{
Funcrion to calculate at-sea observer sampling effort.
}
\details{
NOTE: MPD appears not to include sampled pots that DID NOT have Tanner crab.
As a consequence, the observer effort calculated here DOES NOT INCLUDE sampled pots with
no Tanner crab in them. Uses \code{sqldf::sqldf}.
}
