% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.calcSSsFromZCs.R
\name{adfg.calcSSsFromZCs}
\alias{adfg.calcSSsFromZCs}
\title{Calculate sample sizes for size compositions}
\usage{
adfg.calcSSsFromZCs(dfr, writeCSV = FALSE)
}
\arguments{
\item{dfr}{\itemize{
\item data.frame from call to function \code{adfg.ExtractRawZCsFromMPD}
}}

\item{writeCSV}{\itemize{
\item flag (T/F) to write table to csv file
}}
}
\value{
a dataframe with columns
\itemize{
\item fishery
\item area
\item year
\item sex
\item maturity
\item shell condition
\item ss - sample size
}
}
\description{
Function to calculate sample sizes for size compositions
}
\details{
Uses \code{sqldf::sqldf}.
}
