% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.calcTotalCatchAB.R
\name{adfg.calcTotalCatchAB}
\alias{adfg.calcTotalCatchAB}
\title{Calculate estimated total catch abundance and biomass by expanding observed abundance}
\usage{
adfg.calcTotalCatchAB(tblMPD, tblObsEff, tblTotEff)
}
\arguments{
\item{tblMPD}{- from call to \code{adfg.extractMPD}}

\item{tblObsEff}{- from call to \code{adfg.getAtSeaObserverEffort} or \code{adfg.calcAtSeaObserverEffort}}

\item{tblTotEff}{- from call to \code{adfg.getTotalFisheryEffort}}
}
\value{
tibble with columns
 * fishery
 * area
 * year
 * potlifts - number of pots fished (total effort)
 * measure pots - number of observed pots (sample effort)
 * expFactor - expansion factor
 * sex - "male", "female", "undetermined", "hermaphrodite", or NA
 * variable - "count", "weight", "abundance", or "biomass"
 * value - value of asociated variable
 * type - "observed" or "expanded"
 * units - "ones", "thousands", "kg" or "t"
}
\description{
Function to calculate estimated total catch abundance by expanding observed abundance.
}
\details{
Uses \code{sqldf::sqldf}. Units for 'weight' are kg, for 'abundance' are thousands, and for 'biomass' are t.
}
