% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.calcTotalCatchZCs.R
\name{adfg.calcTotalCatchZCs}
\alias{adfg.calcTotalCatchZCs}
\title{Calculate expanded annual total catch size compositions by crab fishery, area, year, sex and shell condition}
\usage{
adfg.calcTotalCatchZCs(
  tblTotAB,
  tblTotZCsRaw,
  cutpts = seq(from = 5, to = 185, by = 5),
  truncate.low = TRUE,
  truncate.high = FALSE
)
}
\arguments{
\item{tblTotAB}{: dataframe with total catch abundance and biomass by fishery, area, year, sex, and shell condition}

\item{tblTotZCsRaw}{:  data.frame from call to function \code{\link[=adfgRead_MPD]{adfgRead_MPD()}}}

\item{cutpts}{: cutpts for bins}

\item{truncate.low}{: flag to exclude crab with sizes less than minimum cutpt (default=TRUE)}

\item{truncate.high}{: flag to exclude crab with sizes greater than maximum cutpt (default=FALSE)}
}
\value{
a dataframe with columns
\itemize{
\item fishery
\item area
\item sex
\item shell condition
\item year
\item size
\item abundance (in thousands of crab)
}
}
\description{
Function to calculate expanded annual total catch size compositions by crab fishery, area, year, sex and shell condition.
}
\details{
Uses \code{\link[sqldf:sqldf]{sqldf::sqldf()}} and \code{\link[reshape2:cast]{reshape2::dcast()}}. tblTotZCsRaw should have columns
"fishery", "area", "year", "sex", "maturity", "shell condition", "size" and "count".
}
