% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.calcZCsFromMPD.R
\name{adfg.calcZCsFromMPD}
\alias{adfg.calcZCsFromMPD}
\title{Calculate size compositions from measure pot data}
\usage{
adfg.calcZCsFromMPD(dfr, writeCSV = FALSE)
}
\arguments{
\item{dfr}{- data.frame from function \code{adfg.extractMPD}}

\item{writeCSV}{- flag (T/F) to write table to csv file}
}
\value{
a dataframe with columns
* year
* fishery
* area
* sex
* maturity
* shell condition
* size
* count
}
\description{
Function to calculate size compositions from measure pot data.
}
\details{
Uses \code{sqldf::sqldf} and \code{wtsUtilities::Sum}.
}
