% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.ConvertDateToFisheryYear.R
\name{adfg.ConvertDateToFisheryYear}
\alias{adfg.ConvertDateToFisheryYear}
\title{Convert a vector of dates to numeric fishery years}
\usage{
adfg.ConvertDateToFisheryYear(v)
}
\arguments{
\item{v}{- vector of dates as mm-dd-yyyy strings}
}
\value{
numeric vector of same length as v with corresponding fishery years (yyyy of yyyy/yy+1)
}
\description{
Function to convert a vector of dates to numeric fishery years
}
\details{
fishery year extend July 1, yyyy to June 30, yyyy+1
}
