% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.ConvertFisheryNames.R
\name{adfg.ConvertFisheryNames}
\alias{adfg.ConvertFisheryNames}
\title{Convert ADFG fishery names to canonical assessment names}
\usage{
adfg.ConvertFisheryNames(x)
}
\arguments{
\item{x}{- vector of ADFG fishery names to convert canonical assessment names}
}
\value{
character vector with canonical names
}
\description{
Function to convert ADFG fishery names to canonical assessment names.
}
\details{
* original names: "Tanner crab","Tanner E","Tanner W","snow crab","snow","BBRKC","RKC"
* final names:    "TCF",        "TCF",     "TCF",     "SCF",      "SCF", "RKF",  "RKF"
}
