% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.ConvertMaturityCodes.R
\name{adfg.ConvertMaturityCodes}
\alias{adfg.ConvertMaturityCodes}
\title{Convert ADFG measure pot codes to immature/mature/undetermined}
\usage{
adfg.ConvertMaturityCodes(sex)
}
\arguments{
\item{sex}{- vector of sex codes (after conversion to male/female/unknown via \code{adfg.ConvertSexCodes})}
}
\value{
a vector of maturity states (immature/mature/undetermined)
}
\description{
Function to convert ADFG measure pot codes to immature/mature/undetermined.
}
\details{
All male crab have "undetermined" maturity. For now, all females are regarded as
'mature' for purposes of calculating weight (as per B. Daly pers. comm., 2019-04-09.)
}
