% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.ConvertSexCodes.R
\name{adfg.ConvertSexCodes}
\alias{adfg.ConvertSexCodes}
\title{Convert ADFG sex codes to character labels}
\usage{
adfg.ConvertSexCodes(x)
}
\arguments{
\item{x}{- vector of sex codes to convert to character labels}
}
\value{
character vector with labels
}
\description{
Function to convert ADFG sex codes to character labels.
}
\details{
* sex codes: 1,2,3,0
* labels: "male","female","hermaphrodite","undetermined"
}
