% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.ExportWideZCs.R
\name{adfg.ExportWideZCs}
\alias{adfg.ExportWideZCs}
\title{Export size compositions in wide format}
\usage{
adfg.ExportWideZCs(
  dfrSSs,
  dfrZCs,
  byFacs = c("fishery", "area", "sex", "shell condition", "year"),
  sizeCol = "size",
  valCol = "abundance",
  ssCol = "ss",
  csvSSs = "dfrSSs.csv",
  csvZCs = "dfrZCs.csv",
  writeFiles = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{dfrSSs}{\itemize{
\item dataframe with sample size information
}}

\item{dfrZCs}{\itemize{
\item size composition dataframe
}}

\item{byFacs}{\itemize{
\item character vector of size composition factors to include in ouptut
}}

\item{sizeCol}{\itemize{
\item name of column in dfrZCs with sizes
}}

\item{valCol}{\itemize{
\item name of column in dfrZCs with abundance values
}}

\item{ssCol}{\itemize{
\item name of column in dfrSSs with sample sizes
}}

\item{csvSSs}{\itemize{
\item name of csv file to save sample size to
}}

\item{csvZCs}{\itemize{
\item name of csv file to save size compositions to
}}

\item{writeFiles}{\itemize{
\item flag (T/F) to write csv files (default=TRUE)
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
A list with named elements
\itemize{
\item tblZCs - the wide-format size compositions
\item tblSSs - a dataframe with sample sizes
}
}
\description{
Function to export size compositions in wide format.
}
\details{
Uses \code{sqldf::sqldf}, \code{reshape2::dcast}, and \code{wtsUtilities::Sum}.
Input column names should given "as is" and \strong{NOT} be backquoted.
}
