% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.ExtractRawZCsFromMPD.R
\name{adfg.ExtractRawZCsFromMPD}
\alias{adfg.ExtractRawZCsFromMPD}
\title{Get annual total catch size compositions by crab fishery from a csv file}
\usage{
adfg.ExtractRawZCsFromMPD(mpd)
}
\arguments{
\item{mpd}{- dataframe from call to \code{adfg.getMPD}}
}
\value{
a tibble with columns
* fishery
* area
* sex
* maturity
* shell condition
* year
* size
* count
}
\description{
Function to get annual total catch size compositions by crab fishery from a csv file.
}
\details{
Uses \code{readr::read_csv} and \code{stringr::str_sub}.
}
