% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.getAtSeaObserverEffort.R
\name{adfg.getAtSeaObserverEffort}
\alias{adfg.getAtSeaObserverEffort}
\title{Get at-sea observer effort from a csv file}
\usage{
adfg.getAtSeaObserverEffort(
  csv = "AtSeaCrabObserverEffort.1990-2018.csv",
  skip = 4
)
}
\arguments{
\item{csv}{\itemize{
\item filename to read
}}

\item{skip}{\itemize{
\item number of lines to skip (default=4)
}}
}
\value{
tibble with columns:
\itemize{
\item fishery
\item area
\item year - fishery year
\item summary pots - number of summary pots sampled
\item measure pots - number of measure pots
}
}
\description{
Function to get at-sea observer effort from a csv file.
}
\details{
Read a csv file with at-sea observer effort, including sampled pots with no crab.
Uses \code{readr::read_csv}, \code{tibble::add_column}, \code{stringr::str_sub}.
}
