% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.getMPD.R
\name{adfg.getMPD}
\alias{adfg.getMPD}
\title{Extract measure pot data as a tibble from a csv file}
\usage{
adfg.getMPD(csv = "crab_dump-931-v17.csv")
}
\arguments{
\item{csv}{- csv filename with measure pot data}
}
\value{
a tibble
}
\description{
Function to extract measure pot data as a tibble from a csv file.
}
\details{
Uses functions \code{readr::read_csv}, \code{stringr::str_sub}.
}
