% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.getRetainedCatchAB.R
\name{adfg.getRetainedCatchAB}
\alias{adfg.getRetainedCatchAB}
\title{Get annual retained catch abundance and biomass by crab fishery from a csv file}
\usage{
adfg.getRetainedCatchAB(csv, skip = 2)
}
\arguments{
\item{csv}{- the csv file name}

\item{skip}{- number of lines to skip (default=2)}
}
\value{
a tibble with columns
* fishery
* area
* year
* sex
* shell condition
* abundance
* biomass (kg)
}
\description{
Function to get annual retained catch abundance and biomass by crab fishery from a csv file.
}
\details{
Uses \code{readr::read_csv} and \code{stringr::str_sub}.
}
