% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.getTotalFisheryEffort.R
\name{adfg.getTotalFisheryEffort}
\alias{adfg.getTotalFisheryEffort}
\title{Get total fishery effort from a csv file}
\usage{
adfg.getTotalFisheryEffort(csv = "FishingEffortByFleet.1990-2017.csv",
  skip = 2)
}
\arguments{
\item{skip}{- number of lines to skip (default=2)}

\item{-}{the csv file name}
}
\value{
a tibble with columns
* fishery
* area
* year
* potlifts - number of potlifts
}
\description{
Function to get total fishery effort from a csv file.
}
\details{
Uses \code{readr::read_csv}, \code{stringr::str_sub}, \code{dplyr::mutate},
\code{tidyr::gather}, and \code{sqldf::sqldf}.
}
