% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.ScaleAndRebinZCs.R
\name{adfg.ScaleAndRebinZCs}
\alias{adfg.ScaleAndRebinZCs}
\title{Scale and rebin size compositions}
\usage{
adfg.ScaleAndRebinZCs(
  mdfrScl,
  mdfrZCs,
  aggFacs = c("shell condition"),
  scaleCol = "abundance",
  sizeCol = "size",
  valCol = "count",
  cutpts = seq(from = 5, to = 185, by = 5),
  truncate.low = TRUE,
  truncate.high = FALSE,
  returnSampleSizes = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{mdfrScl}{\itemize{
\item dataframe with scaling information
}}

\item{mdfrZCs}{\itemize{
\item "raw" size compositions (counts) to scale and rebin
}}

\item{aggFacs}{\itemize{
\item character vector of biological factors over which to aggregate (e.g., shell condition)
}}

\item{scaleCol}{\itemize{
\item name of column in mdfrScl with scale information
}}

\item{sizeCol}{\itemize{
\item name of column in mdfrZCs with sizes to rebin
}}

\item{valCol}{\itemize{
\item name of column in mdfrZCs with values to scale
}}

\item{cutpts}{\itemize{
\item vector of cutpoints
}}

\item{truncate.low}{\itemize{
\item flag to truncate below first cutpoint (i.e., cutpt\link{1}<-0)
}}

\item{truncate.high}{\itemize{
\item flag to truncate above last cutpoint (i.e., cutpt\link{last}<-Inf)
}}

\item{returnSampleSizes}{\itemize{
\item flag to return sample sizes as well as size compositions (default=FALSE)
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic information
}}
}
\value{
If returnSampleSizes is FALSE, a dataframe with same columns as mdfrZCs, except that the \code{valCol}
column is renamed to that in \code{scaleCol}.

If returnSampleSizes is TRUE, then the return value is a list with named elements
\itemize{
\item tblSS -a dataframe with sample sizes
\item tblZCs - the scaled, rebinned and expanded size compositions
}
}
\description{
Function to scale and rebin size compositions.
}
\details{
Uses \code{sqldf::sqldf}, \code{reshape2::dcast}, \code{wtsUtilities::Sum}, and \code{wtsUtilities::applyCutPts}.
Input column names should given "as is" and \strong{NOT} be backquoted.
}
