% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfgConvert_DateMMDDYYYYtoFisheryYear.R
\name{adfgConvert_DateMMDDYYYYtoFisheryYear}
\alias{adfgConvert_DateMMDDYYYYtoFisheryYear}
\title{Convert a vector of dates as mm-dd-yyyy to numeric fishery years}
\usage{
adfgConvert_DateMMDDYYYYtoFisheryYear(x)
}
\arguments{
\item{x}{- vector of dates as mm-dd-yyyy strings}
}
\value{
numeric vector of fishery years (e.g., 1990 represents 1990/91)
}
\description{
Function to convert a vector of dates as mm-dd-yyyy to numeric fishery years
}
\details{
The fishery year extends from July 1, yyyy to June 30, yyyy+1,so
if mm<7, then year=yyyy-1, otherwise year=yyyy.
}
