% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfgConvert_FisheryNames.R
\name{adfgConvert_FisheryNames}
\alias{adfgConvert_FisheryNames}
\title{Convert ADFG fishery names to canonical assessment names}
\usage{
adfgConvert_FisheryNames(x)
}
\arguments{
\item{x}{\itemize{
\item vector of ADFG fishery names to convert canonical assessment names
}}
}
\value{
character vector with canonical names
}
\description{
Function to convert ADFG fishery names to canonical assessment names.
}
\details{
\itemize{
\item original names: "WBT","EBT","Tanner crab","Tanner E","Tanner W","Tanner","Snow crab","snow crab","Snow","snow","BSSC","BBRKC","RKC"
\item final names:    "TCF","TCF","TCF",        "TCF",     "TCF",     "TCF",   "SCF",      "SCF",      "SCF", "SCF" ,"SCF", "RKF",  "RKF"
}
}
