% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfgConvert_FisheryNamesToAreas.R
\name{adfgConvert_FisheryNamesToAreas}
\alias{adfgConvert_FisheryNamesToAreas}
\title{Convert ADFG fishery names to canonical assessment names}
\usage{
adfgConvert_FisheryNamesToAreas(x)
}
\arguments{
\item{x}{\itemize{
\item vector of ADFG fishery names to convert canonical assessment names
}}
}
\value{
character vector with canonical names
}
\description{
Function to convert ADFG fishery names to canonical assessment names.
}
\details{
The original values are converted to the final values using the following table:
\itemize{
\item{  original:    final    }
\item{WBT:         West 166W}
\item{EBT:         East 166W}
\item{Tanner crab: all EBS  }
\item{ Tanner E:    East 166W}
\item{ Tanner W:    West 166W}
\item{ Tanner:      all EBS  }
\item{ Snow crab:   all EBS  }
\item{ snow crab:   all EBS  }
\item{ Snow:        all EBS  }
\item{ snow:        all EBS  }
\item{ BSSC:        all EBS  }
\item{ BBRKC:       all EBS  }
\item{ RKC:         all EBS  }
}
}
