% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfgConvert_SexCodes.R
\name{adfgConvert_SexCodes}
\alias{adfgConvert_SexCodes}
\title{Convert ADFG sex codes to character labels}
\usage{
adfgConvert_SexCodes(x)
}
\arguments{
\item{x}{\itemize{
\item vector of sex codes to convert to character labels
}}
}
\value{
character vector with labels
}
\description{
Function to convert ADFG sex codes to character labels.
}
\details{
\itemize{
\item sex codes: 1,2,3,0
\item labels: "male","female","hermaphrodite","undetermined"
}

If x is a character vector, values are converted to numeric and elements with non-NA values
are converted using the above substitutions. The elements that generated NA values
(i.e. non-numeric text) are not changed. Elements that can be converted to numeric but
do not match one of the sex codes (e.g., -999) are converted to "undetermined".
}
\examples{
# example code
x = c("1","female","-999",NA);
adfgConvert_SexCodes(x);

}
