% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfgRead_DirectedEffortFile.R
\name{adfgRead_DirectedEffortFile}
\alias{adfgRead_DirectedEffortFile}
\title{Read the directed fishery effort from a csv file}
\usage{
adfgRead_DirectedEffortFile(csv = "FishingEffortByFleet.1990-2017.csv",
  skip = 2, verbose = FALSE)
}
\arguments{
\item{csv}{- the csv file name}

\item{skip}{- number of lines to skip (default=2)}

\item{verbose}{- flag (T/F) to print diagnostic info}
}
\value{
a tibble with columns
* fishery - canonical fishery name
* area - fishery area
* year - crab year
* effort - number of potlifts
}
\description{
Function to read the directed fishery effort from a csv file.
}
\details{
Uses \code{readr::read_csv}, \code{stringr::str_sub}, \code{dplyr::mutate},
\code{tidyr::gather}, and \code{sqldf::sqldf}.
}
