% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfgRead_MPD.R
\name{adfgRead_MPD}
\alias{adfgRead_MPD}
\title{Extract measure pot data as a tibble from a csv file}
\usage{
adfgRead_MPD(csv = "crab_dump-931-v17.csv", date_format = "yyyy-mm-dd")
}
\arguments{
\item{csv}{- csv filename with measure pot data}

\item{date_format}{- string ("yyyy-mm-dd" or "mm-dd-yyyy") indicating date format}
}
\value{
a tibble with columns "fishery","area","EWbySA","EWbyLon","year","fishery_code",
"code_year","trip","adfg","sampdate","spn","statarea","mi_lon","mi_lat",
"spcode","sex","shell","size","legal","count"
}
\description{
Function to extract measure pot data as a tibble from a csv file.
}
\details{
Uses functions \code{readr::read_csv}, \code{stringr::str_sub}.
}
