% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfgRead_SPD.R
\name{adfgRead_SPD}
\alias{adfgRead_SPD}
\title{Extract summary pot data (SPD) as a tibble from a csv file}
\usage{
adfgRead_SPD(
  csv = "TANNER-1990-2020_potsum.csv",
  date_format = "mm-dd-yyyy",
  returnDropped = FALSE
)
}
\arguments{
\item{csv}{\itemize{
\item csv filename with summary pot data
}}

\item{date_format}{\itemize{
\item string ("yyyy-mm-dd" or "mm-dd-yyyy") indicating date format
}}

\item{returnDropped}{\itemize{
\item flag (T/F) to return \emph{dropped} records, not selected records
}}
}
\value{
a tibble with columns "fishery","area","EWbySA","EWbyLon","year","fishery_code",
"code_year","trip","adfg","sampdate","spn","statarea","longitude","latitude",
"female", "sublegal", "legal_ret", "legal_nr", "legal_ur", "tot_legal".
}
\description{
Function to extract summary pot data (SPD) as a tibble from a csv file.
}
\details{
Uses functions \code{\link[readr:read_delim]{readr::read_csv()}}, \code{\link[stringr:str_sub]{stringr::str_sub()}}.
}
\note{
The 'year' values are 'crab year' based on the sample date. The 'code_year' is the
ADFG fishery year based on the fishery code. Prior to rationalization, there may be differences
in these two values.

value columns represent the number of crab in the category.
}
