% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfgRead_TotalCatchABs.R
\name{adfgRead_TotalCatchABs}
\alias{adfgRead_TotalCatchABs}
\title{Get annual total catch abundance and biomass by crab fishery from a 2026+ csv file}
\usage{
adfgRead_TotalCatchABs(csv, skip = 0)
}
\arguments{
\item{csv}{\itemize{
\item the csv file name
}}

\item{skip}{\itemize{
\item number of lines to skip (default=0)
}}
}
\value{
a tibble with columns \cr
\itemize{
\item{fishery}
\item{area}
\item{year}
\item{sex}
\item{maturity}
\item{\verb{shell condition}}
\item{abundance - in numbers of crab}
\item{\code{biomass (kg)} - in kg}
}
}
\description{
Function to get annual total catch abundance and biomass by crab fishery from a 2026+ csv file.
}
\details{
Uses \code{readr::read_csv}, \code{stringr::str_sub}, and \code{sqldf::sqldf}.
}
\note{
The input csv should be in the "total_catch_sex_shell.csv" format provided annually by ADFG
for Tanner crab since 2026. These are the "official" estimates of total catch abundance and biomass
of Tanner crab in the crab fisheries.

\code{shell_condition} may include "undetermined" values, as well as "new shell" and "old shell" values.
}
