% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfgZCs_CalcRetZCsFromDSD.R
\name{adfgZCs_CalcRetZCsFromDSD}
\alias{adfgZCs_CalcRetZCsFromDSD}
\title{Calculate retained catch size compositions from dockside data}
\usage{
adfgZCs_CalcRetZCsFromDSD(dfr, fn = NULL)
}
\arguments{
\item{dfr}{- data.frame from function \code{adfgRead_DSD}}

\item{fn}{- name of file to write csv to (or NULL not to write)}
}
\value{
a dataframe with columns
* year
* fishery
* area
* sex
* maturity
* shell condition
* size
* count
}
\description{
Function to calculate retained catch size compositions from dockside data.
}
\details{
Uses \code{sqldf::sqldf} and \code{wtsUtilities::Sum}.
}
