% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfgZCs_CalcTotZCsFromMPD.R
\name{adfgZCs_CalcTotZCsFromMPD}
\alias{adfgZCs_CalcTotZCsFromMPD}
\title{Calculate size compositions from measure pot data}
\usage{
adfgZCs_CalcTotZCsFromMPD(dfr, fn = NULL)
}
\arguments{
\item{dfr}{\itemize{
\item data.frame from function \code{adfgRead_MPD}
}}

\item{fn}{\itemize{
\item name of file to write csv to (or NULL not to write)
}}
}
\value{
a dataframe with columns
\itemize{
\item year
\item fishery
\item area
\item sex
\item maturity
\item shell condition
\item size
\item count
}
}
\description{
Function to calculate size compositions from measure pot data.
}
\details{
Uses \code{sqldf::sqldf} and \code{wtsUtilities::Sum}.
}
