% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseFisheryCode.R
\name{parseFisheryCode}
\alias{parseFisheryCode}
\title{parse a 4-character ADFG "fishery" code}
\usage{
parseFisheryCode(f)
}
\arguments{
\item{f}{}
}
\value{
dataframe with columns:
* fishery code - f
* fishery
* area
* year
}
\description{
Function to parse 4-character ADFG "fishery" codes.
}
\details{
Uses \code{stringr::sub}, \code{wtsUtilities::substituteValues} functions.
}
