% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.calcSSsFromZCs.R
\name{adfg.calcSSsFromZCs}
\alias{adfg.calcSSsFromZCs}
\title{Calculate sample sizes for size compositions}
\usage{
adfg.calcSSsFromZCs(dfr, writeCSV = FALSE)
}
\arguments{
\item{dfr}{- data.frame from call to function \code{adfg.ExtractRawZCsFromMPD}}

\item{writeCSV}{- flag (T/F) to write table to csv file}
}
\value{
a dataframe with columns
* fishery
* area
* year
* sex
* maturity
* shell condition
* ss - sample size
}
\description{
Function to calculate sample sizes for size compositions
}
\details{
Uses \code{sqldf::sqldf}.
}
