% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.calcTotalCatchAB.R
\name{adfg.calcTotalCatchAB}
\alias{adfg.calcTotalCatchAB}
\title{Calculate estimated total catch abundance and biomass by expanding observed abundance}
\usage{
adfg.calcTotalCatchAB(tblMPD, tblObsEff, tblTotEff)
}
\arguments{
\item{tblMPD}{- from call to \code{adfg.extractMPD}}

\item{tblObsEff}{- from call to \code{adfg.getAtSeaObserverEffort} or \code{adfg.calcAtSeaObserverEffort}}

\item{tblTotEff}{- from call to \code{adfg.getTotalFisheryEffort}}
}
\value{
tibble with columns
 * fishery
 * area
 * year
 * sex
 * totPots - number of pots fished (total effort)
 * obsPots - number of observed pots (sample effort)
 * expFactor - expansion factor
 * obsAbund - observed abdundance
 * obsWgt - individual weight calculated using L-W regresssions
 * totAbund - estimated total catch abundance (millions)
 * totBio - estimated total catch biomass (thousands t)
}
\description{
Function to calculate estimated total catch abundance by expanding observed abundance.
}
\details{
Uses \code{sqldf::sqldf}.
}
