% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.calcTotalCatchAB.R
\name{adfg.calcTotalCatchAB}
\alias{adfg.calcTotalCatchAB}
\title{Calculate estimated total catch abundance and biomass by expanding observed abundance}
\usage{
adfg.calcTotalCatchAB(tblMPD, tblObsEff, tblTotEff)
}
\arguments{
\item{tblMPD}{\itemize{
\item from call to \code{adfg.extractMPD}
}}

\item{tblObsEff}{\itemize{
\item from call to \code{adfg.getAtSeaObserverEffort} or \code{adfg.calcAtSeaObserverEffort}
}}

\item{tblTotEff}{\itemize{
\item from call to \code{adfg.getTotalFisheryEffort}
}}
}
\value{
tibble with columns
\itemize{
\item fishery
\item area
\item year
\item potlifts - number of pots fished (total effort)
\item measure pots - number of observed pots (sample effort)
\item expFactor - expansion factor
\item sex - "male", "female", "undetermined", "hermaphrodite", or NA
\item variable - "count", "weight", "abundance", or "biomass"
\item value - value of asociated variable
\item type - "observed" or "expanded"
\item units - "ones", "thousands", "kg" or "t"
}
}
\description{
Function to calculate estimated total catch abundance by expanding observed abundance.
}
\details{
Uses \code{sqldf::sqldf}. Units for 'weight' are kg, for 'abundance' are thousands, and for 'biomass' are t.
}
\note{
Output from this function should be used as a check on abundance/biomass values
provided by ADFG (obtained from \code{adfg.Read_TotalCatchABs}).
}
