% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.ScaleAndExpandZCs.R,
%   R/adfg.calcTotCatchZCs.R
\name{adfg.calcTotalCatchZCs}
\alias{adfg.calcTotalCatchZCs}
\title{Calculate expanded annual total catch size compositions by crab fishery, area, year, sex and shell condition}
\usage{
adfg.calcTotalCatchZCs(tblTotAB, tblTotZCsRaw, cutpts = seq(from = 5, to
  = 185, by = 5), truncate.low = TRUE, truncate.high = FALSE)

adfg.calcTotalCatchZCs(tblTotAB, tblTotZCsRaw, cutpts = seq(from = 5, to
  = 185, by = 5), truncate.low = TRUE, truncate.high = FALSE)
}
\arguments{
\item{tblTotZCsRaw}{- "raw" total catch size compositions (counts) by fishery, area, year, sex, shell condition and size}

\item{tblRetAB}{- dataframe with total catch abundance and biomass by fishery, area, year, sex, an shell condition}

\item{tblRetAB}{- dataframe with total catch abundance and biomass by fishery, area, year, sex, an shell condition}

\item{tblTotZCsRaw}{- "raw" total catch size compositions (counts) by fishery, area, year, sex, shell condition and size}
}
\value{
a datarame with columns
* fishery
* area
* sex
* shell condition
* year
* size
* abundance (in thousands of crab)

a datarame with columns
* fishery
* area
* sex
* shell condition
* year
* size
* abundance (in thousands of crab)
}
\description{
Function to calculate expanded annual total catch size compositions by crab fishery, area, year, sex and shell condition.

Function to calculate expanded annual total catch size compositions by crab fishery, area, year, sex and shell condition.
}
\details{
Uses \code{sqldf::sqldf} and \code{reshape2::dcast}.

Uses \code{sqldf::sqldf} and \code{reshape2::dcast}.
}
