% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.ScaleAndExpandZCs.R,
%   R/adfg.calcTotalCatchZCs.R
\name{adfg.calcTotalCatchZCs}
\alias{adfg.calcTotalCatchZCs}
\title{Calculate expanded annual total catch size compositions by crab fishery, area, year, sex and shell condition}
\usage{
adfg.calcTotalCatchZCs(
  tblTotAB,
  tblTotZCsRaw,
  cutpts = seq(from = 5, to = 185, by = 5),
  truncate.low = TRUE,
  truncate.high = FALSE
)

adfg.calcTotalCatchZCs(
  tblTotAB,
  tblTotZCsRaw,
  cutpts = seq(from = 5, to = 185, by = 5),
  truncate.low = TRUE,
  truncate.high = FALSE
)
}
\arguments{
\item{tblTotAB}{- dataframe with total catch abundance and biomass by fishery, area, year, sex, and shell condition}

\item{tblTotZCsRaw}{-  data.frame from call to function \code{adfg.getMPD}}

\item{tblRetAB}{- dataframe with total catch abundance and biomass by fishery, area, year, sex, an shell condition}
}
\value{
a datarame with columns
* fishery
* area
* sex
* shell condition
* year
* size
* abundance (in thousands of crab)

a dataframe with columns
* fishery
* area
* sex
* shell condition
* year
* size
* abundance (in thousands of crab)
}
\description{
Function to calculate expanded annual total catch size compositions by crab fishery, area, year, sex and shell condition.

Function to calculate expanded annual total catch size compositions by crab fishery, area, year, sex and shell condition.
}
\details{
Uses \code{sqldf::sqldf} and \code{reshape2::dcast}.

Uses \code{sqldf::sqldf} and \code{reshape2::dcast}. tblTotZCsRaw should have columns
"fishery", "area", "year", "sex", "maturity", "shell condition", "size" and "count".
}
