% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.ExtractRawZCsFromMPD.R
\name{adfg.ExtractRawZCsFromMPD}
\alias{adfg.ExtractRawZCsFromMPD}
\title{Get annual total catch size compositions by crab fishery from a csv file}
\usage{
adfg.ExtractRawZCsFromMPD(mpd)
}
\arguments{
\item{mpd}{\itemize{
\item dataframe from call to \code{adfg.getMPD}
}}
}
\value{
a tibble with columns
\itemize{
\item fishery
\item area
\item sex
\item maturity
\item shell condition
\item year
\item size
\item count
}
}
\description{
Function to get annual total catch size compositions by crab fishery from a csv file.
}
\details{
Uses \code{readr::read_csv} and \code{stringr::str_sub}.
}
