% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.getTotalCatchZCs.R
\name{adfg.getTotalCatchZCs}
\alias{adfg.getTotalCatchZCs}
\title{Get annual total catch size compositions by crab fishery from a csv file}
\usage{
adfg.getTotalCatchZCs(csv, skip = 2, progress = FALSE)
}
\arguments{
\item{csv}{\itemize{
\item the csv file name
}}

\item{skip}{\itemize{
\item number of lines to skip (default=5)
}}

\item{progress}{\itemize{
\item flag to show progress reading csv file
}}
}
\value{
a tibble with columns
\itemize{
\item fishery
\item area
\item sex
\item shell condition
\item year
\item size
\item count
}
}
\description{
Function to get annual total catch size compositions by crab fishery from a csv file.
}
\details{
The input csv file should be one provided by ADFG.
Uses \code{readr::read_csv} and \code{stringr::str_sub}.
}
