% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfg.ScaleAndRebinZCs.R
\name{adfg.ScaleAndRebinZCs}
\alias{adfg.ScaleAndRebinZCs}
\title{Scale and rebin size compositions}
\usage{
adfg.ScaleAndRebinZCs(
  mdfrScl,
  mdfrZCs,
  aggFacs = c("shell condition"),
  scaleCol = "abundance",
  sizeCol = "size",
  valCol = "count",
  cutpts = seq(from = 5, to = 185, by = 5),
  truncate.low = TRUE,
  truncate.high = FALSE,
  returnSampleSizes = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{mdfrScl}{: dataframe with scaling information}

\item{mdfrZCs}{: "raw" size compositions (counts) to scale and rebin}

\item{aggFacs}{: character vector of biological factors over which to aggregate (e.g., shell condition)}

\item{scaleCol}{: name of column in mdfrScl with scale information}

\item{sizeCol}{: name of column in mdfrZCs with sizes to rebin}

\item{valCol}{:  name of column in mdfrZCs with values to scale}

\item{cutpts}{: vector of cutpoints}

\item{truncate.low}{: flag to truncate below first cutpoint (i.e., cutpt[1]<-0)}

\item{truncate.high}{: flag to truncate above last cutpoint (i.e., cutpt[last]<-Inf)}

\item{returnSampleSizes}{: flag to return sample sizes as well as size compositions (default=FALSE)}

\item{verbose}{: flag (T/F) to print diagnostic information}
}
\value{
If \code{returnSampleSizes} is FALSE, a dataframe with same columns as mdfrZCs, except that the \code{valCol}
column is renamed to that in \code{scaleCol}.

If \code{returnSampleSizes} is TRUE, then the return value is a list with named elements
\itemize{
\item tblSS -a dataframe with sample sizes
\item tblZCs - the scaled, rebinned and expanded size compositions
}
}
\description{
Function to scale and rebin size compositions.
}
\details{
Uses \code{\link[sqldf:sqldf]{sqldf::sqldf()}}, \code{\link[reshape2:cast]{reshape2::dcast()}}, \code{\link[wtsUtilities:Sum]{wtsUtilities::Sum()}}, and \code{\link[wtsUtilities:applyCutPts]{wtsUtilities::applyCutPts()}}.
Input column names should given "as is" and \strong{NOT} be backquoted.
}
