% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfgEffort_calcAtSeaObserverEffort.R
\name{adfgEffort_calcAtSeaObserverEffort}
\alias{adfgEffort_calcAtSeaObserverEffort}
\title{Calculate at-sea observer sampling effort from measure pot data}
\usage{
adfgEffort_calcAtSeaObserverEffort(tbl)
}
\arguments{
\item{tbl}{\itemize{
\item measure pot data table from \code{adfg.extractMPD}
}}
}
\value{
dataframe with columns
\itemize{
\item fishery
\item area
\item year
\item numpots
}
}
\description{
Function to calculate at-sea observer sampling effort from measure pot data.
}
\details{
NOTE: MPD appears not to include sampled pots that DID NOT have Tanner crab.
As a consequence, the observer effort calculated here DOES NOT INCLUDE sampled pots with
no Tanner crab in them. Uses \code{sqldf::sqldf}.
}
