% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfgRead_DirectedEffortFile.R
\name{adfgRead_DirectedEffortFile}
\alias{adfgRead_DirectedEffortFile}
\title{Read the directed fishery effort from a csv file}
\usage{
adfgRead_DirectedEffortFile(
  csv = "FishingEffortByFleet.1990-2017.csv",
  skip = 2,
  verbose = FALSE
)
}
\arguments{
\item{csv}{\itemize{
\item the csv file name
}}

\item{skip}{\itemize{
\item number of lines to skip (default=2)
}}

\item{verbose}{\itemize{
\item flag (T/F) to print diagnostic info
}}
}
\value{
a tibble with columns
\itemize{
\item{fishery - canonical fishery name}
\item{area - fishery area}
\item{year - crab year}
\item{effort - number of potlifts}
}
}
\description{
Function to read the directed fishery effort from a csv file.
}
\details{
Uses \code{readr::read_csv}, \code{stringr::str_sub}, \code{dplyr::mutate},
\code{tidyr::gather}, and \code{sqldf::sqldf}.

The input csv file should have columns
\itemize{
\item{fishery year in crab year format (i.e., YYYY/YY)}
\item{area - "E","W", or "all EBS"}
\item{Directed Tanner - effort (number of pots fished)}
\item{Directed Snow   - effort (number of pots fished)}
\item{Directed BBRKC  - effort (number of pots fished)}
}
}
