% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfgRead_MPD.R
\name{adfgRead_MPD}
\alias{adfgRead_MPD}
\title{Extract measure pot data (MPD) as a tibble from a csv file}
\usage{
adfgRead_MPD(csv = "crab_dump.csv", date_format = "yyyy-mm-dd")
}
\arguments{
\item{csv}{\itemize{
\item csv filename with measure pot data
}}

\item{date_format}{\itemize{
\item string ("yyyy-mm-dd" or "mm-dd-yyyy") indicating date format
}}
}
\value{
a tibble with columns "fishery","area","EWbySA","EWbyLon","year","fishery_code",
"code_year","trip","adfg","sampdate","spn","statarea","mi_lon","mi_lat",
"spcode","sex","shell","size","legal","count"
}
\description{
Function to extract measure pot data (MPD) as a tibble from a csv file.
}
\details{
Uses functions \code{\link[readr:read_delim]{readr::read_csv()}}, \code{\link[stringr:str_sub]{stringr::str_sub()}}.
}
\note{
The 'year' values are 'crab year' based on the sample date. The 'code_year' is the
ADFG fishery year based on the fishery code. Prior to rationalization, there may be differences
in these two values.

'count' is the number of measured crab associated with the rest of the row categories.
Unlike the dockside data, this  is 1 for all rows (i.e., one row for each measured crab).
}
