% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfgRead_RetainedCatchABs.R
\name{adfgRead_RetainedCatchABs}
\alias{adfgRead_RetainedCatchABs}
\title{Get annual retained catch abundance and biomass by crab fishery from a csv file}
\usage{
adfgRead_RetainedCatchABs(csv, skip = 2)
}
\arguments{
\item{csv}{\itemize{
\item the csv file name
}}

\item{skip}{\itemize{
\item number of lines to skip (default=2)
}}
}
\value{
a tibble with columns
\itemize{
\item fishery
\item area
\item year
\item sex
\item maturity
\item shell condition
\item abundance
\item biomass (kg)
}
}
\description{
Function to get annual retained catch abundance and biomass by crab fishery from a csv file.
}
\details{
Uses \code{readr::read_csv} and \code{stringr::str_sub}.
}
