% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfgRead_SPD.R
\name{adfgRead_SPD}
\alias{adfgRead_SPD}
\title{Extract summary pot data (SPD) as a tibble from a csv file}
\usage{
adfgRead_SPD(
  csv = "TANNER-1990-2020_potsum.csv",
  date_format = "mm-dd-yyyy",
  col_types = "ccccccddddddddcc"
)
}
\arguments{
\item{csv}{\itemize{
\item csv filename with summary pot data
}}

\item{date_format}{\itemize{
\item string ("yyyy-mm-dd" or "mm-dd-yyyy") indicating date format
}}

\item{col_types}{\itemize{
\item passed to readr::read_csv, specifies column types (dedfault="ccccccddddddddcc")
}}
}
\value{
a tibble with columns "fishery","area","EWbySA","EWbyLon","year","fishery_code",
"code_year","trip","adfg","sampdate","spn","statarea","mi_lon","mi_lat",
"spcode","sex","shell","size","legal","count"
}
\description{
Function to extract summary pot data (SPD) as a tibble from a csv file.
}
\details{
Uses functions \code{\link[readr:read_delim]{readr::read_csv()}}, \code{\link[stringr:str_sub]{stringr::str_sub()}}.
}
\note{
The 'year' values are 'crab year' based on the sample date. The 'code_year' is the
ADFG fishery year based on the fishery code. Prior to rationalization, there may be differences
in these two values.

'count' is the number of crab associated with the rest of the row categories.
}
