% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfgRead_TotalCatchABs.R
\name{adfgRead_TotalCatchABs}
\alias{adfgRead_TotalCatchABs}
\title{Get annual total catch abundance and biomass by crab fishery from a csv file}
\usage{
adfgRead_TotalCatchABs(csv, skip = 5)
}
\arguments{
\item{csv}{- the csv file name}

\item{skip}{- number of lines to skip (default=5)}
}
\value{
a tibble with columns \cr
\itemize{
  \item{fishery}
  \item{area}
  \item{year}
  \item{sex}
  \item{maturity}
  \item{`shell condition`}
  \item{abundance - in numbers of crab}
  \item{`biomass (kg)` - in kg}
}
}
\description{
Function to get annual total catch abundance and biomass by crab fishery from a csv file.
}
\details{
Uses \code{readr::read_csv}, \code{stringr::str_sub}, and \code{sqldf::sqldf}.
}
\note{
The input csv should be the exported "item4a" worksheet from the Excel spreadsheet provided annually by ADFG
for Tanner crab. These are the "official" estimates of total catch abundance and biomass
of Tanner crab in the crab fisheries.
}
