% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfgWrite_TCSAMInputFile.R
\name{adfgWrite_TCSAMInputFile}
\alias{adfgWrite_TCSAMInputFile}
\title{Write ADFG data fo a TCSAM input file}
\usage{
adfgWrite_TCSAMInputFile(
  fishery = NULL,
  fn = "Data.Fishery.ADFG.inp",
  closed = NULL,
  dfrRC_ABs = NULL,
  dfrRC_ZCs = NULL,
  dfrRC_SSs = NULL,
  dfrTC_ABs = NULL,
  dfrTC_ZCs = NULL,
  dfrTC_SSs = NULL,
  dfrEffort = NULL,
  rcCutPts = NULL,
  tcCutPts = NULL,
  likeRC = "NORM2",
  cvRC = 0.05,
  minErrRCa = 100,
  minErrRCb = 100,
  likeTC = "NORM2",
  cvTC = 0.2,
  minErrTCa = 2000,
  minErrTCb = 2000,
  unitsBiomass = "THOUSANDS_MT"
)
}
\arguments{
\item{fishery}{- TCSAM fishery name}

\item{fn}{- output file name}

\item{closed}{- vector of years when fishery was closed}

\item{dfrRC_ABs}{- dataframe with retained catch abundance/biomass data}

\item{dfrRC_ZCs}{- dataframe with retained catch size composition data}

\item{dfrRC_SSs}{- dataframe with retained catch sample size information}

\item{dfrTC_ABs}{- dataframe with total catch abundance/biomass data}

\item{dfrTC_ZCs}{- dataframe with total catch size composition data}

\item{dfrTC_SSs}{- dataframe with total catch sample size information}

\item{dfrEffort}{- dataframe with fishery effort}

\item{rcCutPts}{- vector with cutpts used for retained catch size comps}

\item{tcCutPts}{- vector with cutpts used for total catch size comps}

\item{likeRC}{- likelihood type for retained catch ABs ("NORM2","NORMAL", or "LOGNORMAL")}

\item{cvRC}{- cv for retained catch above min error (default=0.05)}

\item{minErrRCa}{- min error in retained catch abundance (default=100)}

\item{minErrRCb}{- min error in retained catch biomass   (default=100 kg)}

\item{likeTC}{- likelihood type for total catch ABs ("NORM2","NORMAL", or "LOGNORMAL")}

\item{cvTC}{- cv for total catch above min error (default=0.20)}

\item{minErrTCa}{- min error in total catch abundance (default=2000)}

\item{minErrTCb}{- min error in total catch biomass   (default=2000 kg)}

\item{unitsBiomass}{- units for output biomass ("THOUSANDS_MT" [default],"MILLIONS_LBS","KG")}
}
\value{
null
}
\description{
Function to write ADFG data to a TCSAM input file.
}
\details{
None
}
